% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{dateFromWeek}
\alias{dateFromWeek}
\title{Return the start and the end date of a particular week}
\usage{
dateFromWeek(year, wkIndex, wkMethod = "ISO")
}
\arguments{
\item{year}{a value for the year.}

\item{wkIndex}{a value for the week index of the year.}

\item{wkMethod}{a character string for the week-counting method.
Default is "ISO" for ISO-week;
other options are "epiSat", "epiSun" and "epiMon" for epi-week method
defining Saturday, Sunday and Monday as the firstday in a week, respectively.}
}
\value{
a vector of length 2, containing the start and the end date.
}
\description{
This function returns the start and the end date of a particular week.
  The week can be defined by ISO week or epi week.
}
